refactor: Depends on `loophp/iterators`.
