+++ composer.json
@@ -35,7 +35,8 @@
         }
     ],
     "require": {
-        "php": ">= 7.4"
+        "php": ">= 7.4",
+        "loophp/iterators": "^1.3"
     },
     "require-dev": {
         "amphp/parallel-functions": "^1",
