+++ spec/loophp/collection/CollectionSpec.php
@@ -22,8 +22,8 @@ use JsonSerializable;
 use loophp\collection\Collection;
 use loophp\collection\Contract\Collection as CollectionInterface;
 use loophp\collection\Contract\Operation;
-use loophp\collection\Iterator\ClosureIterator;
 use loophp\collection\Operation\AbstractOperation;
+use loophp\iterators\ClosureIterator;
 use OutOfBoundsException;
 use PhpSpec\Exception\Example\FailureException;
 use PhpSpec\Exception\Example\MatcherException;
