+++ src/Operation/Flatten.php
@@ -12,7 +12,7 @@ namespace loophp\collection\Operation;
 use Closure;
 use Generator;
 use Iterator;
-use loophp\collection\Iterator\IterableIterator;
+use loophp\iterators\IterableIterator;
 
 /**
  * @immutable
