+++ src/Operation/Product.php
@@ -13,7 +13,7 @@ use ArrayIterator;
 use Closure;
 use Generator;
 use Iterator;
-use loophp\collection\Iterator\IterableIterator;
+use loophp\iterators\IterableIterator;
 
 /**
  * @immutable
